first setup for opensoar version
