+++ requirements.txt
@@ -1,5 +1,4 @@
-xlwt==0.7.5
+xlwt==1.3.0
 numpy==1.12.0b1
-beautifulsoup==3.2.1
 pyinstaller==3.2
-pygeodesy==17.6.23
+git+https://github.com/glidergeek/opensoar@improvements-from-pysoar
